requirements_frozen.txt
absl-py==2.4.0
accelerate==1.13.0
aeiou==0.0.20
aiobotocore==3.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
alias-free-torch==0.0.6
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
argbind==0.3.9
asttokens==3.0.1
attrs==26.1.0
audioread==3.1.0
auraloss==0.4.0
av==17.0.0
bleach==6.3.0
bokeh==3.9.0
botocore==1.42.70
braceexpand==0.1.7
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
clean-fid==0.1.35
click==8.3.1
clip-anytorch==2.6.0
cloudpickle==3.1.2
colorcet==3.1.0
colorlog==6.10.1
configparser==7.2.0
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
dctorch==0.1.2
decorator==5.2.1
decord==0.6.0
deepspeed==0.18.8
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
docstring_parser==0.17.0
einops==0.8.2
einops-exts==0.0.4
einx==0.4.2
ema-pytorch==0.7.9
encodec==0.1.1
executing==2.2.1
fastapi==0.135.2
fastcore==1.12.31
ffmpy==1.0.0
filelock==3.25.2
fire==0.7.1
flatten-dict==0.4.2
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.46
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
h5py==3.16.0
hf-gradio==0.3.0
hf-xet==1.4.2
hjson==3.1.0
holoviews==1.22.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
hyper-connections==0.4.9
idna==3.11
ImageIO==2.37.3
importlib_metadata==9.0.0
importlib_resources==6.5.2
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonmerge==1.9.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
julius==0.2.7
k-diffusion==0.1.1
kiwisolver==1.5.0
kornia==0.8.2
kornia_rs==0.1.10
laion-clap==1.1.4
lazy-loader==0.5
librosa==0.9.2
lightning-utilities==0.15.3
linkify-it-py==2.1.0
llvmlite==0.46.0
local-attention==1.11.2
Markdown==3.10.2
markdown-it-py==4.0.0
markdown2==2.5.5
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmaudio @ git+https://github.com/hkchengrex/MMAudio.git@974010a026c731054592d8f777218bd9d85a6c24
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
narwhals==2.18.1
networkx==3.6.1
ninja==1.13.0
nitrous_ema==0.0.1
numba==0.64.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch==3.3.0
opencv-python==4.13.0.92
orjson==3.11.7
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pandas==2.3.3
panel==1.8.10
param==2.3.2
parso==0.8.6
pedalboard==0.9.22
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.9.4
plotly==6.6.0
pooch==1.9.0
prefigure==0.0.9
progressbar==2.5
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==3.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.2.0
pynndescent==0.6.0
pyparsing==3.3.2
pypinyin==0.55.0
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytorch-lightning==2.6.1
pytz==2026.1.post1
pyvers==0.2.2
pyviz_comms==3.0.6
PyWavelets==1.9.0
PyYAML==6.0.3
qwen-omni-utils==0.0.9
randomname==0.2.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
resampy==0.4.3
rich==14.3.3
rpds-py==0.30.0
ruff==0.15.7
s3fs==2026.2.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.56.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
SoundFile==0.10.2
soxr==1.0.0
-e git+https://github.com/ZeyueT/AudioX-sync.git@bf5eda95603fc6e88d4345e3a684e88d0795cfae#egg=stable_audio_tools
stack-data==0.6.3
starlette==0.52.1
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.11.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.26
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.6.0+cu124
torch-einops-utils==0.0.30
torch-stoi==0.2.3
torchaudio==2.6.0+cu124
torchcodec==0.11.0
torchdiffeq==0.2.5
torchlibrosa==0.1.0
torchmetrics==1.9.0
torchsde==0.2.6
torchvision==0.21.0+cu124
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
trampoline==0.1.2
transformers==5.3.0
triton==3.2.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
umap-learn==0.5.11
urllib3==2.6.3
uvicorn==0.42.0
v-diffusion-pytorch==0.0.2
vector-quantize-pytorch==1.28.0
wandb==0.25.1
wcwidth==0.6.0
webdataset==1.0.2
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.7
wget==3.2
wrapt==2.1.2
x-transformers==1.26.6
xyzservices==2025.11.0
yarl==1.23.0
zipp==3.23.0
